Fix (most) MeasureImageOverlap errors

+++ cellprofiler/modules/measureimageoverlap.py
@@ -508,8 +508,8 @@ the two images. Set this setting to “No” to assess no penalty."""
 
         returns a tuple of the Rand Index and the adjusted Rand Index
         """
-        ground_truth_labels = ground_truth_labels[mask].astype(numpy.uint64)
-        test_labels = test_labels[mask].astype(numpy.uint64)
+        ground_truth_labels = ground_truth_labels[mask].astype(numpy.uint32)
+        test_labels = test_labels[mask].astype(numpy.uint32)
         if len(test_labels) > 0:
             #
             # Create a sparse matrix of the pixel labels in each of the sets
@@ -518,7 +518,7 @@ the two images. Set this setting to “No” to assess no penalty."""
             # labeled with label I in the ground truth and label J in the
             # test set.
             #
-            N_ij = scipy.sparse.coo_matrix((numpy.ones(len(test_labels)),
+            N_ij = scipy.sparse.coo_matrix((numpy.ones(len(test_labels), dtype=numpy.uint32),
                                             (ground_truth_labels, test_labels))).toarray()
 
             def choose2(x):
